+++ src/Core/Batch.php
@@ -125,11 +125,13 @@ class Batch implements BatchOperationsInterface
                 'additionalParameters' => $additionalParameters,
             ]
         );
+        
+        $useFieldsInsteadOfId = $apiMethod === 'crm.address.delete';
 
         try {
             $this->clearCommands();
             foreach ($entityItemId as $cnt => $itemId) {
-                if (!is_int($itemId)) {
+                if (!$useFieldsInsteadOfId && !is_int($itemId)) {
                     throw new InvalidArgumentException(
                         sprintf(
                             'invalid type «%s» of entity id «%s» at position %s, entity id must be integer type',
@@ -140,7 +142,8 @@ class Batch implements BatchOperationsInterface
                     );
                 }
 
-                $parameters = ['ID' => $itemId];
+                $parameters = $useFieldsInsteadOfId ? ['fields' => $itemId] : ['ID' => $itemId];
+                // TODO: delete after migration to RestAPI v2
                 if ($apiMethod === 'entity.item.delete') {
                     $parameters = array_merge($parameters, $additionalParameters);
                 }
@@ -201,8 +204,11 @@ class Batch implements BatchOperationsInterface
 
         try {
             $this->clearCommands();
+            
+            $useFieldsInsteadOfId = $apiMethod === 'crm.address.update';
+            
             foreach ($entityItems as $entityItemId => $entityItem) {
-                if (!is_int($entityItemId)) {
+                if (!$useFieldsInsteadOfId && !is_int($entityItemId)) {
                     throw new InvalidArgumentException(
                         sprintf(
                             'invalid type «%s» of entity id «%s», entity id must be integer type',
@@ -219,10 +225,17 @@ class Batch implements BatchOperationsInterface
                         );
                     }
 
-                    $cmdArguments = [
-                        'id' => $entityItemId,
-                        'fields' => $entityItem['fields']
-                    ];
+                    if ($useFieldsInsteadOfId) {
+                        $cmdArguments = [
+                            'fields' => $entityItem['fields']
+                        ];
+                    } else {
+                        $cmdArguments = [
+                            'id' => $entityItemId,
+                            'fields' => $entityItem['fields']
+                        ];
+                    }
+
                     if (array_key_exists('params', $entityItem)) {
                         $cmdArguments['params'] = $entityItem['params'];
                     }